Config Dialog: removed redundant RestoreCurrentConfig();
This is redundant because h/w options are never applied to live variable settings directly.
They are only stored to the registry (after the user has accepted) and only from there are they applied to live variable settings.

So resetting the live variable settings is not required as they have not been touched (yet).